Run association analysis

to be written. calls PLINK. (include specifying pheno_names from simulation: GWAS use pheno, for LTFH use LTFH_pheno, GWAX use GWAX_pheno) NOTE: creates a temporary .bed file if bed is FALSE. This takes some additional time. Also: File extension will be .assoc or .qassoc depending on the used model.

analysis_association(
  geno_file,
  pheno_file,
  pheno_name,
  out_file,
  bed = TRUE,
  plink_path = TRUE
)

Arguments

geno_file
string specifying path to genotypes file, including file extension.

pheno_file
string specifying path to phenotypes file, including file extension.

pheno_name
column name of phenotype to be used in analysis.

out_file
string specifying path and name of the resulting output file, excluding file extension.

bed
logical indicating whether or not the genotypes file is a .bed file. FALSE if the file is .ped.

plink_path
TRUE if user has added PLINK to the system variable "Path". Otherwise a string specifying the path to PLINK.

Value

Does not return anything, but PLINK writes out_file to disk, containing results of the performed analysis. The columns in the printed dataset depends on which phenotype is used for the analysis.