family_simulation

Simulate genotypes for individuals with family history, where each individual has a fixed number of siblings. Parents' genotypes are simulated and used for simulating the genotypes of individuals and their siblings. family_simulation makes use of parallel computation in order to decrease the running time.

family_simulation(n, m, q, hsq, k, path = "", sib = 0)

Arguments

n
number of genotypes (individuals).

m
number of SNPS per genotype.

q
number of causal SNPs, i.e. SNPs that effect chances of having the phenotype.

hsq
squared heritability parameter.

k
prevalence of phenotype.

path
directory where the files will be stored. If nothing is specified, family_simulation writes its files in the current working directory.

sib
number of siblings per individual.

Value

Does not return any value, but prints the following five files to the path parameter specified in the function call:

Three text files:

beta.txt - a file of m rows with one column. The i'th row is the true effect of the i'th SNP.

MAFs.txt - a file of m rows with one column. The i'th row is the true Minor Allelle Frequence of the i'th SNP.

phenotypes.txt - a file of n rows, number of columns depend on number of siblings. The file contains the phenotype and liability of each individual as well as information on the liabilities and phenotype status of their parents and siblings.

genotypes.map - a file created such that PLINK will work with the genotype data.

genotypes.ped - the simulated genotypes in a PLINK-readable format.

Warning

Simulating large datasets lead to very large files (eg. in the order of 100k genotypes with 100k SNPs takes up approximately 40GB of space). Please ensure that you have sufficient disk space at the disk where path resides prior to running the simulation. See FUNCTION-NAME DOCUMENTATION REFERENCE(this is a placeholder until function is implemented) in order to convert the generated .ped to the smaller .bed format.