Multivariate normal distributions

Simulate random values from a multivariate normal distribution.

rmnorm(n, S)

Arguments

n
number of random vectors that is to be simulated.

S
a valid covariance matrix.

Value

Returns a matrix with n rows and i columns, where i is determined by the size of the specified covariance matrix.

Examples

rmnorm(n = 10, S = covmatrix(sib = 0, hsq = 0.5))

#>               [,1]       [,2]        [,3]        [,4]
#>  [1,] -0.989980265 -1.3815049 -0.05707177  0.30842814
#>  [2,]  0.180536421  0.6252939  0.42977802  0.20516351
#>  [3,] -1.723405627 -0.2632125 -2.08199156 -0.46176695
#>  [4,] -0.003939495 -1.1572232  0.68815681  1.40255493
#>  [5,]  0.439504144  0.8018447  1.98628668  0.07111709
#>  [6,]  0.812049634 -0.5052984  0.31487327  0.29498934
#>  [7,] -1.288219622 -1.6573382 -1.51951485 -2.28744932
#>  [8,] -0.174885398 -0.2120806 -0.10236285 -0.34383470
#>  [9,] -0.172675198  0.2112812 -0.85972785 -0.26604786
#> [10,] -0.199902940 -0.8462514 -1.51467180  1.09028652