Simulate genotypes in the R session

A quick way of simulating a small amount of data to be stored in the R-session. Note that this function only simulates data for individuals without family history, i.e. they are not simulated as offspring. Hence, the simulation cannot be used in a study of family history.

testsim(n, m, q, hsq, k, to_ped = T)

Arguments

n
number of genotypes (individuals).

m
number of SNPS per genotype.

q
number of causal SNPs, i.e. SNPs that effect chances of having the phenotype.

hsq
squared heritability parameter.

k
prevalence of phenotype.

to_ped
TRUE/FALSE indicating if the output should be in a PLINK-friendly format or not.

Value

Returns a list of 4 entries, containing a matrix of genotypes, a vector specifying indexes of causal SNPs, a vector the liability of individual i, and a vector specifying whether individual i has the phenotype.

Examples

testsim(10, 15, 5, 0.5, 0.05, TRUE)

#> $genotypes
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,] "1"  "1"  "1"  "1"  "1"  "1"  "a"  "a"  "c"  "a"   "a"   "a"   "a"  
#>  [2,] "2"  "1"  "1"  "1"  "1"  "1"  "c"  "a"  "c"  "c"   "a"   "a"   "a"  
#>  [3,] "3"  "1"  "1"  "1"  "1"  "1"  "a"  "a"  "a"  "a"   "c"   "a"   "a"  
#>  [4,] "4"  "1"  "1"  "1"  "1"  "1"  "a"  "a"  "c"  "c"   "a"   "a"   "a"  
#>  [5,] "5"  "1"  "1"  "1"  "1"  "1"  "a"  "a"  "c"  "c"   "c"   "a"   "a"  
#>  [6,] "6"  "1"  "1"  "1"  "1"  "1"  "a"  "a"  "c"  "a"   "c"   "a"   "a"  
#>  [7,] "7"  "1"  "1"  "1"  "1"  "1"  "a"  "a"  "c"  "c"   "c"   "a"   "a"  
#>  [8,] "8"  "1"  "1"  "1"  "1"  "1"  "a"  "a"  "a"  "a"   "c"   "a"   "a"  
#>  [9,] "9"  "1"  "1"  "1"  "1"  "1"  "a"  "a"  "a"  "a"   "a"   "a"   "a"  
#> [10,] "10" "1"  "1"  "1"  "1"  "1"  "a"  "a"  "a"  "a"   "c"   "a"   "c"  
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,] "a"   "a"   "a"   "c"   "c"   "a"   "a"   "a"   "a"   "c"   "a"   "c"  
#>  [2,] "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "c"   "a"   "a"  
#>  [3,] "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "c"   "a"   "a"  
#>  [4,] "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "c"   "c"   "c"  
#>  [5,] "a"   "a"   "a"   "c"   "a"   "a"   "a"   "a"   "a"   "c"   "a"   "c"  
#>  [6,] "a"   "a"   "a"   "a"   "a"   "c"   "a"   "c"   "c"   "c"   "a"   "a"  
#>  [7,] "a"   "a"   "a"   "c"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "c"  
#>  [8,] "a"   "c"   "a"   "c"   "a"   "a"   "a"   "c"   "a"   "a"   "a"   "c"  
#>  [9,] "a"   "a"   "a"   "a"   "a"   "a"   "a"   "c"   "a"   "c"   "c"   "c"  
#> [10,] "a"   "a"   "a"   "a"   "a"   "c"   "c"   "a"   "a"   "c"   "a"   "c"  
#>       [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36]
#>  [1,] "a"   "c"   "a"   "a"   "a"   "a"   "a"   "c"   "a"   "c"   "a"  
#>  [2,] "a"   "c"   "a"   "a"   "a"   "a"   "a"   "c"   "a"   "a"   "a"  
#>  [3,] "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"  
#>  [4,] "a"   "a"   "a"   "a"   "a"   "a"   "a"   "c"   "c"   "a"   "a"  
#>  [5,] "a"   "c"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"  
#>  [6,] "a"   "c"   "a"   "a"   "a"   "a"   "a"   "c"   "a"   "a"   "a"  
#>  [7,] "a"   "c"   "c"   "c"   "a"   "a"   "a"   "a"   "a"   "c"   "a"  
#>  [8,] "a"   "a"   "a"   "c"   "a"   "a"   "a"   "c"   "a"   "a"   "a"  
#>  [9,] "a"   "c"   "c"   "a"   "a"   "a"   "a"   "c"   "c"   "c"   "a"  
#> [10,] "a"   "c"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"   "a"  
#> 
#> $`SNP-Turn`
#>             [,1]
#>  [1,]  0.0000000
#>  [2,] -0.5157641
#>  [3,]  0.0000000
#>  [4,]  0.0000000
#>  [5,]  0.0000000
#>  [6,]  0.0000000
#>  [7,]  0.1620436
#>  [8,]  0.1989016
#>  [9,]  0.0000000
#> [10,]  0.0000000
#> [11,] -0.5891360
#> [12,]  0.0000000
#> [13,]  0.6530182
#> [14,]  0.0000000
#> [15,]  0.0000000
#> 
#> $liability
#>              [,1]
#>  [1,]  0.09900456
#>  [2,] -0.65255031
#>  [3,]  0.90321239
#>  [4,] -0.20884569
#>  [5,] -0.21119547
#>  [6,]  0.15145020
#>  [7,] -1.56643033
#>  [8,]  2.25616669
#>  [9,] -0.56772078
#> [10,]  0.34484495
#> 
#> $Phenotype
#>  [1] 1 1 1 1 1 1 1 2 1 1
#>